added another chest and key
I was able to successfully add in another chest with a key to unlock it but I can't seem to figure out how to make the keys open up the chests separately. I think it has something to do with them being the same but I don't know how to fix it.