bartable and shelf made + uv mapped
all my main assets are finished except for three but that's only because I haven't added color to them and put them into my main scene. Afterwards it will just be making a few bottles and updating the walls and floor to match the environment and I'm pretty much finished.